Cfdp 90 twilio notify support (#3)
* Testing Push Notification registration with local Prividers

* Updated production env to point to cf apollos api

* Created Storybook story for each of the necessary on boarding screens

* Onboarding Landing Page

Created an on boarding landing page that requests a user enters their phone number or email, validates the input, and runs code specific to the data entered

* Onboarding Passcode verification

Created the form for logging in with either an SMS code or username/password. When creating an instance of the Passcode component, pass in the property "type" of values "code" or "password" for specific language and verification endpoint.

* Onboarding Profile Information

Created a page for onboarding that asks for First Name, Last Name, Gender, and Birthdate. It then updates the Person record upon submitting. First Name and Last Name are required, while the others are optional

* Onboarding Enable Notifications

Screen made to enable notifications from. Include a "Skip for now" secondary call to action underneath the enable

* Removed unecessary directory

* Identity onboarding screen

* Passcode onboarding screen

* Profile Information onboarding screen

* Removed directory for renaming

* Enable Notifications onboarding screen

* updated onboarding slide references

* updated Onboarding stack navigator

* Updated Auth provider

* updated tabar icons to be primary

* Updated logout logic

* Updated main App Stack Navigator to account for new Onboarding flow