Project Migration
* Migrated the project over to React 16.3.1 and React Native 0.63.2
* Migrated the project to React Navigation 5.x
* Migrated the project to @apollo/client
* Updated Xcode proejct settings

At this point in time, the app loads successfully and you can interact with Onboarding as expected and land on the Home Tab.

The theming is all over the place for Onboarding and no content currently loads for the Home Feed. An API update is required to get that content and I'll be working on that next.